requirements.txt
mypy==1.5.1
black
biopython==1.84
